competition phase start
